requirements.txt
 ##SJSU CMPE 138 Fall 2021 TEAM8
 flask
 flask-mysql
